% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseStepwise.R
\name{analyseStepwise}
\alias{analyseStepwise}
\title{Summarize stepwise output}
\usage{
analyseStepwise(stepwise.summary)
}
\arguments{
\item{stepwise.summary}{list}
}
\value{
FIX ME
}
\description{
FIX ME
}
