% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD:man/centralities_network.Rd
% Please edit documentation in R/centralities_network.R
\name{centralities_network}
\alias{centralities_network}
\title{Calculate network centralities}
\usage{
centralities_network(
=======
% Please edit documentation in R/network_centralities.R
\name{network_centralities}
\alias{network_centralities}
\title{Calculate network centralities}
\usage{
network_centralities(
>>>>>>> 42f1e3243b0b136dd72c76edb43e40e46dfee4bb:man/network.centralities.Rd
  network = g,
  include.distance = FALSE,
  distance.matrix = NA,
  distance.att = NA
)
}
\arguments{
\item{network}{An igraph graph.}

\item{distMatrix}{matrix. Default is false. If a distance matrix is provided the function calculates the median distance of movements.}
}
\value{
return a vector listing network centralities
}
\description{
calculate basic network centralities
}
\examples{
<<<<<<< HEAD:man/centralities_network.Rd
centralities_network()
=======
network_centralities()
>>>>>>> 42f1e3243b0b136dd72c76edb43e40e46dfee4bb:man/network.centralities.Rd

}
