% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_network.R
\name{generate_network}
\alias{generate_network}
\title{Generate network}
\usage{
generate_network(
  movements,
  premises,
  splitByEdges = NA,
  splitByVertex = NA,
  save.network = TRUE,
  path.save = getwd(),
  prefixe = "network",
  remove.duplicated.edges = F,
  remove.loop.edges = T
)
}
\arguments{
\item{movements}{A dataframe listing the movements defined by at least an origin ('from' column) and a destination ('to' column).}

\item{premises}{A dataframe listing the origin and destination sites defined by at least the site identifier (1st column) and the nodal attributes.}

\item{splitByEdges}{FIX ME - A vector or characters including one or several time periods among year, semester and month. FALSE is leading to the production of the complete network.}

\item{splitByVertex}{FIX ME - A vector or characters including one or several time periods among year, semester and month. FALSE is leading to the production of the complete network.}

\item{save.network}{logical. Default is TRUE. Produce an '.Rda' file for each generated network and save it under the declared path and prefixe.}

\item{path.save}{String defining the location file to save the networks.}

\item{prefixe}{String defining the prefixe if output files.}
}
\value{
return a list of networks
}
\description{
generate network and subnetworks
}
\examples{
data(listPremises)
data(listMovements)
generate_network(movements = listMovements, premises = listPremises, save.network = FALSE)


}
