% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoDistanceInMetresMatrix.R
\name{GeoDistanceInMetresMatrix}
\alias{GeoDistanceInMetresMatrix}
\title{Add time variables}
\usage{
GeoDistanceInMetresMatrix(df.geopoints, siteID = "siteID")
}
\arguments{
\item{df}{FIX ME}

\item{formatDate}{FIX ME}
}
\value{
FIX ME
}
\description{
FIX ME - by Peter Rosenmai on 30 Jan 2014 https://eurekastatistics.com/calculating-a-distance-matrix-for-geographic-points-using-r/
}
\examples{
sites <- data.frame(siteID = letters[1:3], lat = c(5.1,2.2,3.4), lon = c(25.1,72.2,53.4))
GeoDistanceInMetresMatrix(sites)

}
