% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.network.R
\name{map.network}
\alias{map.network}
\title{Map network}
\usage{
map.network(
  network,
  country,
  premisescolours = NA,
  path.rasters = "",
  title = "network map",
  topfeature = "point",
  path.color = "grey",
  path.size = 0.03,
  point.size.coef = 6
)
}
\arguments{
\item{network}{An igraph graph including at least longitude, latitude and type attributes. Expected names for those attributes are respectively lon, lat and type.}

\item{country}{Vector including country names and the ISO codes used by the function getData of raster package to download mapbase from 'GADM'.}

\item{path.rasters}{Path to save rasters of mapbase.}

\item{title}{Title of the map}

\item{topfeature}{String defining the top feature on map: path (edges) or point (nodes).}

\item{path.color}{String defining path color.}

\item{path.size}{String defining path size.}

\item{point.size.coef}{Numerical. Defining the upper scale for point size. Default is 6.}
}
\value{
return a plot
}
\description{
plot network on a map
}
\examples{
map.network()

}
