% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movements}
\alias{movements}
\title{movements}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
\item{\code{from}}{double origin site identification}
\item{\code{to}}{double destination site identification}
\item{\code{MType}}{double type of movement (loading: 'C' or unloading: 'D')}
\item{\code{year}}{double year of the event}
\item{\code{semester}}{double semester of the event}
\item{\code{semester}}{double month of the  event}
}
}
\usage{
movements
}
\description{
Fictive movements dataset
}
\keyword{datasets}
