% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.centralities.R
\name{network.centralities}
\alias{network.centralities}
\title{Calculate network centralities}
\usage{
network.centralities(
  network = g,
  include.distance = FALSE,
  distance.matrix = NA,
  distance.att = NA
)
}
\arguments{
\item{network}{An igraph graph.}

\item{distMatrix}{matrix. Default is false. If a distance matrix is provided the function calculates the median distance of movements.}
}
\value{
return a vector listing network centralities
}
\description{
calculate basic network centralities
}
\examples{
network.centralities()

}
