% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectedERGM.R
\name{selectedERGM}
\alias{selectedERGM}
\title{Summarize stepwise output}
\usage{
selectedERGM(
  stepwise.summary,
  g,
  outputpath = "",
  distMatrix = NA,
  nsimGof = 2000,
  nSimNet = 5,
  maxtimeERGM = 60 * 60 * 2,
  country.plot = c("FR"),
  nrepAIC = 3,
  verbose = T
)
}
\arguments{
\item{stepwise.summary}{FIX ME}

\item{g}{FIX ME}

\item{outputpath}{FIX ME}

\item{distMatrix}{FIX ME}

\item{nsimGof}{FIX ME}

\item{nSimNet}{FIX ME}

\item{maxtimeERGM}{FIX ME}

\item{country.plot}{FIX ME}

\item{nrepAIC}{FIX ME}
}
\value{
FIX ME
}
\description{
FIX ME
}
