% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{sites}
\format{
A data frame with 200 rows and 10 variables:
\describe{
\item{\code{siteID}}{double identification of premises}
\item{\code{farmGroup}}{double identification of company}
\item{\code{type}}{double type of premise}
\item{\code{freerange}}{double does the premise include freerange reasing}
\item{\code{BRS}}{double Batch rearing system of the premise}
\item{\code{lon}}{double Longitudinal coordinates}
\item{\code{nbRP}}{double Maximal number of reproductive sows in the farm}
\item{\code{nbPW}}{double Maximal number of growing pigs in the farm}
\item{\code{nbFat}}{double Maximal number of barrows in the farm}
}
}
\usage{
sites
}
\description{
Fictive premises dataset
}
\keyword{datasets}
