% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swineNetwork.representation.R
\name{swineNetwork.representation}
\alias{swineNetwork.representation}
\title{Plot a representation of network}
\usage{
swineNetwork.representation(
  networktoplot = g,
  title = NA,
  movements.type.lab = "animType",
  premises.type.lab = "type"
)
}
\arguments{
\item{title}{Title of the map}

\item{movements.type.lab}{string. Default is "animType"}

\item{premises.type.lab}{string. Default is "type"}

\item{network}{An igraph graph including at least longitude, latitude and type attributes. Expected names for those attributes are respectively lon, lat and type.}

\item{edge.attr}{string. Default is NULL}
}
\value{
return a plot
}
\description{
plot network on a map
}
\examples{
FIX ME

}
