% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timewindows.R
\name{timewindows}
\alias{timewindows}
\title{Add time variables}
\usage{
timewindows(df, formatDate = "\%d/\%m/\%Y")
}
\arguments{
\item{df}{FIX ME}

\item{formatDate}{FIX ME}
}
\value{
FIX ME
}
\description{
FIX ME
}
\examples{
data(listMovements)
timewindows(listMovements, formatDate = "\%Y-\%m-\%d") \%>\% head
}
