% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{sites}
\format{
A data frame with 200 rows and 10 variables:
\describe{
\item{\code{siteID}}{double, identification of locations}
\item{\code{company}}{double, identification of company}
\item{\code{type}}{double, type of farm ("FA": farrow,  "FF": farrow-to-finish,  "FI": finish,
"FPW": farrow-to-postwean,  "MU": multiplier,  "NU": nucleus,  "PW":postwean,  "PWF": postwean-to-finish)}
\item{\code{outdoor}}{logical, does the site include outdour rearing system?}
\item{\code{BRS}}{double, Batch rearing system of the site}
\item{\code{lon}}{double, Longitudinal coordinates}
\item{\code{lat}}{double, Latitude coordinates}
\item{\code{nbRP}}{double, Maximal number of reproductive sows in the farm}
\item{\code{nbPW}}{double, Maximal number of growing pigs in the farm}
\item{\code{nbFat}}{double, Maximal number of barrows in the farm}
}
}
\usage{
sites
}
\description{
Fictive sites dataset
}
\keyword{datasets}
