% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise4ERGM.R
\name{stepwise4ERGM}
\alias{stepwise4ERGM}
\title{Stepwise selection for ergm}
\usage{
stepwise4ERGM(
  base.formula = NULL,
  nbworkers = 10,
  elements = NA,
  network.name,
  results.path = "../data/ERGM_output/combinations/",
  maxduration = 60 * 60 * 2,
  g,
  distMatrix = NA,
  mode = "forward",
  stepwise.summary = list(),
  baseline.aic = NA,
  predict.limit = TRUE,
  verbose = F
)
}
\arguments{
\item{base.formula}{string}

\item{nbworkers}{integer}

\item{elements}{list of terms}

\item{network.name}{string}

\item{results.path}{string}

\item{maxduration}{integer/ Duration in seconds}

\item{g}{network}

\item{distMatrix}{matrix}

\item{mode}{String. "forward" or "backward"}

\item{stepwise.summary}{list}

\item{baseline.aic}{integer}

\item{verbose}{logical}
}
\value{
Saved objects
}
\description{
Process to forward, backward or bidirectional stepwise selection of ergm based on term
}
